- Removed useless window class. - Removed useless logging to validate event_manager creation and destruction. - Changed SYSid to be a strongly-typed enum. Cast it to an uint when core logs errors. - Calculate frame time in the main loop, and report frames per second every 2 seconds.

+++ engine/core.cpp
@@ -20,7 +20,7 @@ bool core::sys_exists(SYSid sid)
 void core::add_sys(SYSid sid, sys_ptr sys)
 {
 	if(sys_exists(sid)) {
-		throw std::runtime_error(MKSTR("System with ID " << sid << " is already in the systems vector!"));
+		throw std::runtime_error(MKSTR("System with ID " << uint(sid) << " is already in the systems vector!"));
 	}
 	
 	m_systems[sid] = sys;
@@ -28,7 +28,7 @@ void core::add_sys(SYSid sid, sys_ptr sys)
 sys_ptr core::get_sys(SYSid sid)
 {
 	if(!sys_exists(sid)) {
-		throw std::runtime_error(MKSTR("System with ID " << sid << " does not exist, cannot retrieve!"));
+		throw std::runtime_error(MKSTR("System with ID " << uint(sid) << " does not exist, cannot retrieve!"));
 	}
 	
 	return m_systems[sid];
